advance refactoring typography
